requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
Pillow==9.1.1
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
